Add Azure CosmosDB MongoDB vCore option for agent memory
This is pretty much an exact duplication of the current teachability code, except it uses MongoDB vCore instead of ChromaDB.  Why?  Because ChromaDB stores all its information in memory, so it is not ideal for long conversations or for containers/vms that delete everything on restart.  This allows memories to be stored and recalled via a permanent database rather than an ephemeral one, and it is not stored in memory on the virtual machine so it will not slow down the execution of the agent workflow.