Update utils.py
Seems `_kwargs.update(kwargs)` will cause errors when using the Azure OpenAI service. It forces the model set in the environment to change to "gpt-4o", which can cause issues (e.g., `The API deployment for this resource does not exist.`). 

The model name in Azure can be customized. For example, I created a model based on gpt-4o, but its name is `my-gpt-4o`, and this name should be used in the API call.