fix: Add deferral to RegisterAgentType and (Add/Remove)Subscription
Unlike with the InProcessRuntime, there is a two-phase initialization, first when AgentsApp is built (when initial agents are registered) and when it StartAsync()s and connects to the Gateway. Unfortunately, it is possible to attempt to send direct RPC calls to the Gateway before the message channel is opened; in this case, the Gateway has no connected client corresponding to the RPC's clientId, and falls over.

The fix is to defer registering agents and subscriptions with the gateway until after the connection is established after .StartAsync() is called.