fix: Race condition between channel open and RPC
Right now we rely on opening the channel to associate a ClientId with an entry on the gateway side. This causes a race when the channel is being opened in the background while an RPC (e.g. MyAgent.register()) is invoked.

If the RPC is processed first, the gateway rejects it due to "invalid" clientId.