fix: Recurrent SendMessageAsync deadlock in message handler
In the .NET InProcessRuntime we tried to minimize the amount of tasks created. Unfortunately, this results in a deadlock when a send message handler is attempting to SendMessage during the handling of the incoming message.

The fix is to create a new task for delivering the message in the message processing loop, which creates a new logical stack to run the delivery, freeing the loop to process the next delivery request.

* Also fixes passing exceptions and cancellations back to the waiting task.
* Also fixes a build slowdown on Windows due to uv and llama-cpp