ci: Remove --locked from uv sync in Integration test project (#5993)
For some reason --locked is causing the CI to fall over in the codesign
pipeline.

Unclear why it is not happening on the GitHub Actions side - though it
may be an artifact of the way we have the uv cache set up (and the uv
install version does not match between the two). Getting to a building
state, then will investigate further.